FP32-f16c mode performance improvement

+++ src/el_def.hpp
@@ -45,7 +45,9 @@ template <typename... Types> struct ConvImplTypes {
 };
 
 namespace conv_impl {
-  using FP16 = ConvImplTypes<short, short, short, short>;
+  // TarrayTypes: FP32/FP16/FP32_F16
+  // GarrayTypes: FP32/FP16/INT8_F32/INT8_F16
+  using FP16 = ConvImplTypes<float, short, short, short>;
   using FP32 = ConvImplTypes<float, float, float, float>;
   using FP32_F16 = ConvImplTypes<float, float, short, float>;
   using INT8_F16 = ConvImplTypes<uint8_t, int8_t, short, float>;
